reform: Optimisation of code structure.
